+++ src/LIN3S/AdminBundle/Extension/ListField/ActionsListFieldType.php
@@ -65,13 +65,13 @@ final class ActionsListFieldType implements ListFieldType
     /**
      * {@inheritdoc}
      */
-    public function header($options, Entity $configuration)
+    public function header($name, Entity $configuration)
     {
-        if (!isset($options['name'])) {
-            throw new \InvalidArgumentException('Name key is required');
+        if (!isset($name)) {
+            throw new \InvalidArgumentException('Field to be rendered must be passed as string');
         }
 
-        return $this->translator->trans($options['name']);
+        return $this->translator->trans($name);
     }
 
     /**