+++ CHANGELOG.md
@@ -6,6 +6,7 @@ To get the diff for a specific change, go to https://github.com/LIN3S/AdminBundl
 To get the diff between two versions, go to https://github.com/LIN3S/AdminBundle/compare/v0.2.0...v0.3.0
 
 * 0.5.0
+    * Added fr translation messages
     * Refactored and unified registries
     * [BC break] Made configuration, dependency injection, extension, registry final.
     * [BC break] `lin3s_admin.action`, `lin3s_admin.list_field` and `lin3s_admin.list_filter` tags now require an `alias`