+++ CHANGELOG.md
@@ -6,6 +6,8 @@ To get the diff for a specific change, go to https://github.com/LIN3S/AdminBundl
 To get the diff between two versions, go to https://github.com/LIN3S/AdminBundle/compare/v0.2.0...v0.3.0
 
 * 0.5.0
+    * [BC break] Changed LIN3SAdminBundle to Lin3sAdminBundle so, be caution with `lin3_s`
+    configuration, now it is `lin3s_admin`
     * Added fr translation messages
     * Refactored and unified registries
     * [BC break] Made configuration, dependency injection, extension, registry final.